pyproject.toml
[project]
name = "munger-matics"
version = "0.1.0"
description = "Personal accounting, budgeting and financial planning application"
readme = "README.md"
authors = [
    { name = "Ranajoy Banerji", email = "14226544+ranajoy-cosmo@users.noreply.github.com" }
]
requires-python = ">=3.12"
dependencies = [
    "streamlit>=1.45.0",
    "polars>=1.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "duckdb>=1.5.1",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.9.0",
    "mypy>=1.0.0",
    "mkdocs-material>=9.0.0",
    "pre-commit>=4.0.0",
]
flows = [
    "prefect>=3.0.0",
]

[project.scripts]
munger-matics = "munger_matics:main"

[build-system]
requires = ["uv_build>=0.8.13,<0.12.0"]
build-backend = "uv_build"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.coverage.run]
source = ["src/munger_matics"]

[tool.coverage.report]
show_missing = true

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true
warn_unused_ignores = true
plugins = ["pydantic.mypy"]
# Gradually tighten as the codebase matures:
# check_untyped_defs = true
# warn_return_any = true
# disallow_untyped_defs = true
